Ensure ParamRef resolves synchronous generators asynchronously
